refactor : MemberResponseUtils로 구조 개선
