+++ src/main/java/site/codemonster/comon/domain/auth/controller/MemberController.java
@@ -5,6 +5,7 @@ import site.codemonster.comon.domain.auth.dto.request.MemberProfileUpdateRequest
 import site.codemonster.comon.domain.auth.dto.response.MemberInfoResponse;
 import site.codemonster.comon.domain.auth.entity.Member;
 import site.codemonster.comon.domain.auth.service.MemberService;
+import site.codemonster.comon.domain.auth.utils.MemberResponseUtils;
 import site.codemonster.comon.domain.team.dto.response.TeamAbstractResponse;
 import site.codemonster.comon.domain.teamMember.entity.TeamMember;
 import site.codemonster.comon.domain.teamMember.service.TeamMemberService;
@@ -32,6 +33,7 @@ public class MemberController {
 
     private final MemberService memberService;
     private final TeamMemberService teamMemberService;
+    private final MemberResponseUtils memberResponseUtils;
 
     @PostMapping
     public ResponseEntity<ApiResponse<?>> createMemberProfile(
@@ -54,7 +56,7 @@ public class MemberController {
 
         return ResponseEntity.status(HttpStatus.CREATED)
                 .contentType(MediaType.APPLICATION_JSON)
-                .body(ApiResponse.successResponseWithData(memberService.getMemberProfileResponse(updatedMember)));
+                .body(ApiResponse.successResponseWithData(memberResponseUtils.getMemberProfileResponse(updatedMember)));
     }
 
     @GetMapping("/own-profile") // 자신의 회원 프로필 조회
@@ -62,7 +64,7 @@ public class MemberController {
 
         return ResponseEntity.status(HttpStatus.OK)
                 .contentType(MediaType.APPLICATION_JSON)
-                .body(ApiResponse.successResponseWithData(memberService.getMemberProfileResponse(member)));
+                .body(ApiResponse.successResponseWithData(memberResponseUtils.getMemberProfileResponse(member)));
     }
 
     @GetMapping("/profile/{uuid}") // uuid로 회원 프로필 조회
@@ -72,7 +74,7 @@ public class MemberController {
 
         return ResponseEntity.status(HttpStatus.OK)
                 .contentType(MediaType.APPLICATION_JSON)
-                .body(ApiResponse.successResponseWithData(memberService.getMemberProfileResponse(findMember)));
+                .body(ApiResponse.successResponseWithData(memberResponseUtils.getMemberProfileResponse(findMember)));
     }
 
 
@@ -86,7 +88,7 @@ public class MemberController {
                 .map(TeamAbstractResponse::of)
                 .toList();
 
-        MemberInfoResponse memberInfoResponse = memberService.getMemberInfoResponse(member, teamAbstractResponses);
+        MemberInfoResponse memberInfoResponse = memberResponseUtils.getMemberInfoResponse(member, teamAbstractResponses);
 
         return ResponseEntity.status(HttpStatus.OK)
                 .contentType(MediaType.APPLICATION_JSON)
